Notes on Solution to Travis Bugs
There were several problems going on:

- My yml document was formatted incorrectly.  First level does not
need dashes.

- BUT the Travis website interface was confusing.  I couldn't see the full
error through a click.  Instead I discovered by accident you could see the error
by hovering over the message.

- Also on the main page for a repository, there's a differnece between
"Restart build" in front of you on the queue (which is linked to a
particular commit) and
going to "More options" and clicking "Trigger Build." The latter will
trigger a build of the lastes commit, while the former will simply
do a rebuild of the commit cited in the queue.

- There were some troubles with Travis's new website too and their
migration of legacy repositories from Git, but that's seemed to have
too many glitches/misturns to describe.  Not sure what was going on
there.

But at least problem seems solved for now and I know more about the
Travis website!